+++ tests/api/ky/sense/api-ky-sense-role_test.py
@@ -8,22 +8,7 @@ Unit Tests (using pytest) for:
 
 import pytest
 
-from tests.api.ky.sense import (
-    _assert_sense_title1,
-    _assert_sense_title2,
-    _assert_zero_diff1,
-    _assert_zero_diff2,
-    _assert_empty_pls1,
-    _assert_empty_pls2,
-    _assert_role_title,
-    _assert_role1,
-    _assert_role2,
-    _assert_role3,
-    _assert_role4,
-    _assert_role5,
-    _assert_role6,
-    _assert_role7,
-)
+from tests.api.ky.sense import *
 
 # Pytest fixtures  #############################################################
 
@@ -44,63 +29,52 @@ class TestRole:  # =============================================================
     # title  -------------------------------------------------------------------
 
     def test_sense_title1(_, opt):
-        print(opt)
-        _assert_sense_title1(opt)
+        assert_sense_title1(opt)
 
     def test_sense_title2(_, opt):
-        print(opt)
-        _assert_sense_title2(opt)
+        assert_sense_title2(opt)
 
     # role  --------------------------------------------------------------------
 
     def test_sense_role_title1(_, opt):
-        print(opt)
-        _assert_role_title(opt)
+        assert_role_title(opt)
 
     def test1(_, opt):
-        print(opt)
-        _assert_role1(opt)
+        assert_role1(opt)
 
     def test2(_, opt):
-        print(opt)
-        _assert_role2(opt)
+        assert_role2(opt)
 
     def test3(_, opt):
-        print(opt)
-        _assert_role3(opt)
+        assert_role3(opt)
 
     def test4(_, opt):
-        print(opt)
-        _assert_role4(opt)
+        assert_role4(opt)
 
     def test5(_, opt):
-        print(opt)
-        _assert_role5(opt)
+        assert_role5(opt)
 
     def test6(_, opt):
-        print(opt)
-        _assert_role6(opt)
+        assert_role6(opt)
 
     def test7(_, opt):
-        print(opt)
-        _assert_role7(opt)
+        assert_role7(opt)
+
+    def test_all_roles(_, opt):
+        assert_all_roles(opt)
 
     # zero diff  ---------------------------------------------------------------
 
     def test_zero_diff1(_, opt):
-        print(opt)
-        _assert_zero_diff1(opt)
+        assert_zero_diff1(opt)
 
     def test_zero_diff2(_, opt):
-        print(opt)
-        _assert_zero_diff2(opt)
+        assert_zero_diff2(opt)
 
     # empty PLs  ---------------------------------------------------------------
 
     def test_empty_pls1(_, opt):
-        print(opt)
-        _assert_empty_pls1(opt)
+        assert_empty_pls1(opt)
 
     def test_empty_pls2(_, opt):
-        print(opt)
-        _assert_empty_pls2(opt)
+        assert_empty_pls2(opt)